feat: 기능 구현

+++ adapters/rest/src/main/java/com/hwanhee/bloggy/adapters/rest/dto/MetaDto.java
@@ -1,16 +1,16 @@
 package com.hwanhee.bloggy.adapters.rest.dto;
 
-import com.hwanhee.bloggy.domain.blog.Meta;
+import com.hwanhee.bloggy.domain.model.Meta;
 import lombok.Value;
 
 @Value(staticConstructor = "of")
 public class MetaDto {
-    int totalCount;
-    int pageableCount;
+    Long totalCount;
+    Long pageableCount;
     boolean isEnd;
 
     public static MetaDto empty() {
-        return new MetaDto(0, 0, false);
+        return new MetaDto(0L, 0L, false);
     }
 
     public static MetaDto from(Meta meta) {
